[Kernels] Add support for head_dim = 80 on Hopper and Blackwell MHA
Swizzle 128B assumes that the contiguous dimension in shared memory is a
multiple of 128 bytes. For bf16 data types, this constrains the depth to
be a multiple of 64. This PR introduces `padded_depth` for the shared
memory physical dimension, ensuring it is always a multiple of 128
bytes. As a result, the depth requirement is generalized to just be a
multiple of 16. In particular, this PR adds support for `head_dim = 80`
on H100 and B200.

MODULAR_ORIG_COMMIT_REV_ID: 0169d2ba57b307947d47ad9f1e9b0d8e87f7f810