[Mojo] Unbind Parameter Values in Struct Generator Op (#68129)
The struct generator op represents a parameterized version of the
closure struct. The parameter of the struct generator op contains the
captured parameter values. The kgen.closure attribute is a value that
represents the value bound to this parameter. It should not occur in the
struct.generator.op. This patch corrects that.

MODULAR_ORIG_COMMIT_REV_ID: 8ccb84b09b38487da2608674ad629010d2e62e0f