[Mojo] Report Type Mismatches as Pass Failure (#68240)
If the type of the synthesized function does not match the type of the
closure attribute symbol, fail the pass. This should never be a user
facing error because it is a compiler bug if we hit it.

MODULAR_ORIG_COMMIT_REV_ID: 508ebc53d2ac23af894bba0e8efc02597e470740