[mojo-parser] mangle param decls when constructing trait default method wrapper (#68310)
Fixes MOCO-2379.

When creating the wrapper for a defaulted trait method inside a struct,
the parameter decls of the trait method need to be mangled in the
context of the struct. We essentially end up with:
```
trait T:
    fn foo[B: Barable](self) -> B:
        return B()

struct S[B: Int](T):
    # synthesized wrapper:
    fn foo[B': Barable](self) -> B':
        return T.foo[B'](self)
```

Note that only the param decls of the FnOp for the wrapper use mangled
params. The metadata for the parameters in the FnTypeGeneratorType still
mentions the unmangled names, since they only serve as metadata, and are
not referenced or load bearing in any capacity.

MODULAR_ORIG_COMMIT_REV_ID: dae5f9536ba42d838dc924cd3c454847c8a027c1