[stdlib] `Variant` and `Optional` auto-optimizes on type trivialities.
Hello,
this is early bottom-up work for auto-optimizing.

The PR is status: draft,
because i'm not sure if this is the way to go, is it?

What we have here is quite encouraging.
Mojo's meta-programing is powerful,
we can auto-optimize the types based on their parameters.
(Even with *Ts)

So for example if T is trivial on an Optional,
the optional can become Trivial too for that dunder.

This is bottom-up, because Optional uses Variant[_NoneType, T].
The Optional itself ask the Variant for its trivialities.
And the Variant get it from all T in Ts.

So if we implement this for InlineArray too,
InlineArray[Optional[Int]] would get its trivialities like this:
_NoneType&Int -> Variant[_NoneType, Int] -> Optional -> InlineArray.

Any ideas and thoughts?

InlineArray trivialities could work in an nested way too:
InlineArray[InlineArray[Int, 64], 64]

But not sure if it is not "too soon" for theses features,
should we wait for using T.__*__is_trivial ?