[mojo-lang] Expand DeclResolution bookkeeping for better recursion diag. (#68461)
This expands the DeclResolver to keep the current stack of in-process
resolving declarations.  This allows us to know what the current decl
is and allows us to produce a stack trace back up to the decl that is
causing a cyclic resolution problem (if we wanted to).

This information allows us to print a stack trace of notes showing
HOW a decl get recursively referenced, not just saying "it was".

MODULAR_ORIG_COMMIT_REV_ID: a4e13cdb89d85f7cb63bee868879be508b8b878c