[KGEN] [NFC] When parsing an MLIR attribute from String, remove unneeded copy of the string. This was a workaround and it was fixed in upstream. (#68478)
NFC

`mlir::parseAttribute` was fixed in upstream to always make a copy of
the input, so it now works for non-null-terminated inputs. No need to
make a copy of the input in our codebase before calling it.

MODULAR_ORIG_COMMIT_REV_ID: e76a9cef45565ba619eabbc5129deb88c02816f6