[mojo-lang] Use the ASTPrinter to print parameter names. (#68730)
This moves off using a bespoke and weird printer for parameter names
to using the (getting better) ASTPrinter interfaces. This requires being
in the right scope to print correctly, so it introduces some infra for
it.

The benefit of this is we get out of the "Parameter #12" regime, and
more
into a "this is inferred parameter x.y.z" which is a lot more user
friendly.

MODULAR_ORIG_COMMIT_REV_ID: ca86f136c2c888532b7d3982e5c85bedc8ad6b5a