[Kernels] Do not parameterize the rank for allgather
This is a compilation time optimization since we do not use rank as a
parameter within the kernel and as a result there is no benefit for
parameterizing the allgather on the rank. This means that we do not
compile the kernel ngpu times and instead compile the kernel once.

MODULAR_ORIG_COMMIT_REV_ID: cf1abc958b45ed452f48ea38adf91570f2ca1123