[mojo-parser] Fix func type conversion clarifying parameter indexing issue (#68758)
When adding clarifying parameters to the front of the parameter list,
all existing index param refs need to be shifted backwards. This was
initially done correctly when calculating the new thunk input param
types, but when creating UnboundAttrs at the end for binding onto the
symbol constant, the type of the UnboundAttrs were calculated using the
unshifted param types. This simply changes them to use the shifted param
types.

This error only occurs when the type being rebound contains strong
requirements for what it contain, e.g. !lit.origin requires its mut attr
to be i1 type, and dies immediately if it's not. But if this was instead
a wrapper Origin type, we do not check at replacement time if the
replaced binding is still valid, and since it ultimately doesn't matter,
it didn't end up causing problems.

MODULAR_ORIG_COMMIT_REV_ID: 5cd8fe2f5474df3e1478c5f857248d4b44c50301