[KGEN] Update KGEN::getModularEnvAttr() to properly store bool values as UnitAttr in EnvAttr (#68655)
`KGEN::EnvAttr` is a wrapper over `DictionaryAttr`. There are two
functions that create `EnvAttr`. One function (`parseDefines`) assumes
that boolean values are stored as `UnitAttr`, the other
(`getModularEnvAttr`) assumes those are stored as `BoolAttr`. When the
latter one is called with a bool input, `EnvAttr::validator()` asserts.

Changes:

* [Bug fix] Changed `getModularEnvAttr` to create UnitAttr attribute for
boolean True, and to create nothing for boolean False. This is
consistent with the documented spec of `EnvAttr`, and this fixes MOCO-2166.
* [NFC] Moved the code that queries a value in `EnvAttr` from
`IREvaluator.cpp` into a method of `EnvAttr` itself. This way, the
querying can be unit-tested.
* Added unit test for both methods that create `EnvAttr`: (1) via
`parseDefines`, (2) via `getModularEnvAttr`
* Added unit test for querying a value in `EnvAttr`.

The end result should be that `EnvAttr` is more self-contained now, and
its constructors have uniform approach on how to represent boolean
values.

The change is somewhat controversial. The alternative approach (which I
implemented first but rolled back) would be to support `BoolAttr` inside
`EnvAttr`, and remove `UnitAttr`. Basically, make it so that "absence of
a name in the dictionary" is not the same as "name=false". The advantage
of that alternative approach is that one could then enumerate all the
defines in a environment, whereas the current approach cannot
distinguish boolean False from no attribute at all, so enumerate is not
possible.

I went with the "UnitAttr" approach because this changes less code. I
added unit tests for all the methods of `EnvAttr`, so it would be much
easier to change the approach in the future.

MAX_PYTHON_ORIG_REV_ID: acd72a9cf5169d23f56a0b6b0509b8fb0b3154cc