[mojo-lang] Unify struct parameter names + function param names
This fixes a problem where flattening the parameter list of a method
tossed away the parameter names of the enclosing struct, leading to
error messages like `failed to infer parameter '' ...` when the
inferred parameter was on the struct.

Fixing this uncovered a latent problem - Mojo was allowing functions
to define parameters with the same name as their enclosing structs,
which soft-of-kind-of-worked in some cases, but didn't work in general.

This patch moves to preserving the names correctly in the flattened
parameter list. We still need to directly reject this with an error
message
(tracked by MOCO-2420).

This patch completes a set of work that finally eliminates parameter
inference failures that refer to parameter number. We should now always
complain about a named parameter, or a named parameter of an
argument or param.

MAX_PYTHON_ORIG_REV_ID: 33ab531e489f9a1235554bb2b3756aed2218138a