[mojo-lang] tidy up copying logic even more. (#68907)
Explicit copying isn't really a thing in the mojo language, it is
just a specific method that can be synthesized.  Pull the logic for
handling that case up into the synth logic for the method, instead of
IREmitter haven't to know about it as a concept.

MODULAR_ORIG_COMMIT_REV_ID: 2f09a88af575f292f8d727d85147114c8ffc3ea8