[mojo-lang] Rewrite note processing in param inference diagnostics. (#68913)
This rewrites processing of "explainer notes" in the parameter inference
machinery to attach the problem to the original error, rather than
emitting
highly redundant note as a second diagnostic.  This also fixes the
"use rebind" hint to be a note like it should be, separated from the
diagnosis
of the problem.

While this does improve the quality of the diagnostics very slightly,
there is no major progression here: this is mostly an internal cleanup
and simplification of the implementation.

MODULAR_ORIG_COMMIT_REV_ID: 0288154d76eb07f0a32ffb758139928de3adf2cf