[mojo-lang] Rework 'mut' argument writeback logic to fix bugs (#68929)
This reworks writeback support to fix a couple of problems and simplify
it.
We now support more flexible/aggressive copy elision as shown in the
`subscript_assignment_inplace` examples, and don't generate incorrect
IR by storing through an immutable ref returned by getitem as shown in
`subscript_assignment_writeback`.

The basic observation here is that Mojo /must/ use a setitem call when
a computed lvalue has both a getitem and setitem and the 'got' value
isn't required. This is important for Dict in particular, because the
setitem
creates a new slot, but getitem aborts if the entry doesn't exist even
if
it can return a mutable ref.

This is progress, but still not fully correct, the FIXME'd example in
`subscript_assignment_inplace` is still not calling the setter like it
should.

MODULAR_ORIG_COMMIT_REV_ID: 105f520505ce2dcfeb018ceb2a3bc64688098512