[Init] Always print stack  traces in non-prod builds (#69064)
We never set this CI env var, even on CI because of how bazel is strict
with what env vars make it in. Instead we can just print this always for
local dev. For production builds printing this would likely be useless
since there won't be symbols, and we rely on sentry for that case.

Whether this stacktrace is useful is still dependent on llvm-symbolizer
being available, but that is for many targets automatically.

MODULAR_ORIG_COMMIT_REV_ID: 55ca74eb341688d67ffbe3e0d42ec41a055ceaa8