[kgen] LowerSemanticCF calling raising func does not fallthru (#69098)
`doesFallThrough` determines what happens at the end of a block. Calling
a raising function may cause it to raise, but does not mean that at the
end of the surrounding block, we'll fall through, since there may be a
return or a set of always-returning if-branches.

MODULAR_ORIG_COMMIT_REV_ID: 7f6e4190e71affa3c4ff6f84e977fc134cba40fd