[Pipelines] Fix dupe served_model_name argument
We define `--served-model-name` twice, once in model config and once in
the cli args. This leads to click warnings that we're registering the
same argument twice (though only in the nightly/rc for some reason)

The fix here is to just remove the dupe definition. A

MODULAR_ORIG_COMMIT_REV_ID: 030ef619444fa6811ab27204d94d635aaaa8d902