[Scheduler] Hotfix issue where it appeared that we were scheduling 0 sized batches
There was a bug where it appeared that we were scheduling 0 sized
batches.

What actually occurred is that we scheduled a CE req with BS=1 with
chunked
prefill enabled. During handling of chunked prefill, we remove the
request
from the input batch. This makes it appear that the batch is now empty.

To fix this, we remove the code in handling chunked prefill that mutates
input. This requires some additional refactoring to make it so that we
do not add partially encoded requests to TG batch.

Lets land this hotfix and do some cleanup afterwards. This is a bit
ugly.

MODULAR_ORIG_COMMIT_REV_ID: 20905609b8ba8adac2a00ae78053bb8dc2183df3