[Benchmarking] Replace nvitop with new max.diagnostics package
nvitop's shutdown procedure is buggy and causes crashes on shutdown.
This is causing problems for us. Also, it's mixed-license which is
annoying to deal with compliance-wise. It also doesn't support AMD.

Rip out nvitop and replace it with something made just for us: a new
max.diagnostics package.

The max.diagnostics package offers an abstraction over NVML and RSMI,
allowing accessing GPU information of both NVIDIA and AMD GPUs. It also
offers a background collector that runs in a separate process, so it's
not affected by the GIL.

`benchmark_serving.py` is updated to use this new package instead of
nvitop.

MAX_TESTS_TESTS_ORIG_REV_ID: 4d0e485621723e830ab2895a4675747e4922e365