[Mojo] Do Not Rely On Symbol Contents To Infer Convention (#69076)
Before this patch the contents of the MemSymbolTriple were used to
determine if a value was capture by move or capture by copy. Now that we
are lifting the assumption that closures are not copyable we need a way
to store all three symbols and yet still capture by move. This patch
accomplishes that by adding a UnitAttr to the MemSymbolTriple to
indicate if the value should be moved or copied into the closure.

MODULAR_ORIG_COMMIT_REV_ID: d7622fc854d78ed1e87e552fb3fb9e302149dd71