[Context] Fix Edge Case with Chunked Prefill
This fixes a rare data dependent edge case with Chunked Prefill and
pre-emptions.
Assume we have a single text context with prompt_len=10 tokens and
chunk_size=4, that generates new_tokens=5, before being pre-empted.
During pre-emption, we call reset, setting the start_idx=0 to reprocess
the entire prompt. While chunking, we run into the scenario in which the
start_idx=12 (chunk 1, start_idx=0; chunk 2, start_idx=4; chunk 3,
start_idx=8; chunk 4, start_idx=12), with end_idx=15 (prompt_len +
new_tokens). In this scenario, `needs_ce=False` incorrectly, as
start_idx > prompt_len. This leads to the scenario in which we returned
output to the server, without any new tokens. Previously this silently
passed to the server without concern. With the change to finish_reason,
we are now validating this data, raising an error when this occurs.

To solve for this in the future we likely want to do some of the
following:
- We may want to introduce a GenerationStatus.CHUNKED to make this more
  explcit.
- We may want to change the TextGenerationOutput object, to return None
  instead of an empty token array, when completion_start_idx ==
completion_end_idx. This eliminates the invalid state early and would
ensure consistent behaviour is enforced downstream.
- Add tests to ensure that we drive out edges cases for invalid context
  state.
MODULAR_ORIG_COMMIT_REV_ID: 8702e00fc2ed12e22c565bbcdc733b82903897c6