[stdlib] Introduce `UnsafePointer.swap_pointees`.
This adds a function to swap potentially aliasing pointers.
This optimizes the case when `T.__moveinit__is_trivial` is `True` by
avoiding the branch when checking if the two pointers are equal.

This updates `Span` and `List`'s `swap` function to utilize this instead
of needing to manually check/assert that the two passed in indexes are
not equal.

MODULAR_ORIG_COMMIT_REV_ID: ad0b5b72127dde43a5609e7fb6885a6f05b8b522