[mojo-lang] Optimize sugar rep (#69925)
This changes SugarAttr to not maintain the entire operand tree in the
sugared form - only maintain the sugar in the sugar state. The
observation
here is that the canonical form will remove all sugar, the "original"
form
will be the thing that underlies the sugar (but is primarily sugared),
and
the result is the union. As such we can reduce nesting which reduces the
size of the MLIR that we have to look at in the crazy cases.

MODULAR_ORIG_COMMIT_REV_ID: da0ba19328a21c3c2ee987498819f7bb243cf822