[mojo-lang] Fixes bug with traitful struct extensions (#70009)
See the `alias` line in the new test case.

Basically, this is triggering struct body resolution, which is trying to
figure out whether it inherits AnyType and other traits, which causes it
to call doesNominalTypeConformTo, which looks for all extensions in
scope, which panics because the extensions arent resolved yet and we're
accessing their canonicalTrait which is empty.

Fix: Make it so doesNominalTypeConformTo triggers signature resolution
of extensions before it accesses their canonicalTrait.

But then another bug arises:

- Resolve alias ZScalar = ZSIMD[...]
- Body resolve struct ZSIMD
- conformsToTrait checks extensions
- Trigger extension signature resolution <-- This step is new, this is
Fix 1
- Extension parses inheritance using *structAstDecl (ZSIMD) as scope <--
NEXT BUG
- Look up "ZConvertibleToPython" in ZSIMD scope <-- VERY WEIRD
- SharedState::lookupAndResolveDecl asks to body-resolve struct ZSIMD <-
CYCLE

The fix to that one is to parse the inheritance list using something
else as its scope. We should either:
- Resolve the trait the perspective of the extension itself
- Resolve the trait from the perspective of the extension's parent (file
scope)

This PR currently does the latter. I suspect the former might be better
though. (Thoughts?)

**PRs that depend on this:**
- #70029 (wi2)
  - #70031 (wi3)
    - #70042 (wi4)
      - #70045 (se-simdconstruct)
        - #70046 (se-simdall) (final PR)
      - #69889 (se-z) (final PR)

MODULAR_ORIG_COMMIT_REV_ID: 6d61b89448099e6672e7f5545d734d5023ca520b