[stdlib] fix leak/OOB read in StackTrace
StackTrace calls KGEN_CompilerRT_GetStackTrace to obtain the stack trace
and this function returns a pointer to a string buffer along with that
buffer's size. The buffer is _NOT_ nul terminated. Unfortunately
StackTrace was assuming it _WAS_ nul terminated, and was reading off the
end of the buffer as a result. We fix this by nul terminating our buffer
before we copy the stack-trace in.

Additionally StackTrace stored a ArcPointer[UnsafePointer[UInt8]] which
was never freed. We instead store ArcPointer[OwnedPointer[UInt8]]. Note
that we need to set this owned pointer to null, which is against the
contract of OwnedPointer. We can't use Optional here because Optional
requires T: Copyable. It looks like OwnedPointer stores things in a
plain UnsafePointer so the UB here doesn't make it through to MLIR/LLVM.

Ideally StackTrace would store a mojo String, instead of
ArcPointer[OwnedPointer[UInt8]], but StackTrace needs to be
@register_passable, which String is not.

MODULAR_ORIG_COMMIT_REV_ID: b3ad4a504f7fb7d5b0341205fd68329890aa58e2