[Stdlib] Add ability to specify fast math flags to fma op
This PR adds the ability to specify fast math flags for the fused
multiply-add (fma) operation in Mojo's SIMD type. The implementation
introduces a FastMathFlag enum that allows developers to control
mathematical optimization behaviors for the fma operation.

We do not export the FastMathFlag enum as a public type since this is a
more advanced feature that should not be used lightly.

MODULAR_ORIG_COMMIT_REV_ID: 63deb143a330d33091ac06cf890cabaa9d41f0d1