[mojo-lang][QoI] Remove Sugarizer from attr canonicalization, ~NFC (#70282)
This removes a well intended IR optimization that I added to shrink the
size of SugarAttrs.  The problem is that later param substitution needs
the canonical form in a sugar expr to simplify things.  For example,
having `apply(Bool.and, x, y)` inside a sugar expr wouldn't get
simplified
when x and y get turned into `{1}, {1}` because the canonical form isn't
there to be simplified.

This optimization breaks downstream things and is a bunch of complexity,
so just remove it.

MODULAR_ORIG_COMMIT_REV_ID: 5f0857cc13657d1b004fffe83b64525dafadb793