[mojo-lang] Improve Sugar safety in origins and other things. (#70291)
As we continuously increase the amount of sugar that can exist
in the Mojo type system, we need to adapt new things to look through
it. This is an incremental step forward, that will be heavily tested
in future patches.

MODULAR_ORIG_COMMIT_REV_ID: 310e2be32cc6cd37fbcedf29f873afb30783ae48