[kernels] fix a edge cases of prefix_product calculation
This is a very subtile bug appeared in one of our migrate from NDBuffer
to LayoutTensor. This problem cases a 4D layout to squash into a 3D one.

The problem is with our `_prefix_product2` function where we call the
`is_tuple` function where we check if the tuple is not a value. The
`is_value()` function requires the int tuple to have len of 1 and also
the value to be greater then -65534. This check would fail if we have
some huge value in the tuple and we multiply this by -1, making the
value a tuple when it is in fact a value.

I just simple add more constrain in the `_prefix_product2` function so
that value get clamped to a `UNKNOWN_VALUE`.

MODULAR_ORIG_COMMIT_REV_ID: 81fd781405434204400c8d1ffab384c750e51b02