[stdlib] Fix incorrect implicit casting for `UnsafePointer` & `LayoutTensor` P7.
This updates the `store` function for both `UnsafePointer` and
`LayoutTensor` to require `mut=True`. This is required as if the
underlying pointer is _not_ mutable, then the `store` operations should
not be enabled - as this could lead to data races or other buggy code.

MODULAR_ORIG_COMMIT_REV_ID: ce40a6915ed3766e2651e1e1fb8520fa44d4b29c