[Serve] [CI] Add missing dependency for HF tests (#71274)
A recent refactoring (in #71115) removed pipelines depending on
max.serve, which is how it should be. These tests were erroneously
relying on that dependency without declaring it, and broke. Declare the
dependency explicitly, so that it's sure to be available.

Slack thread:
[Internal link]

MAX_INTEGRATION_TESTS_REV_ID: 0bb95866d569849bc64766ea8ecdb3946c9e54c4