[Models] Remove custom sampling params per model for smoke test (#71285)
The smoke tests aim to be as simple as possible. The goal
is to verify that we get similar results as vLLM and SGLang for our
models. We generally want to avoid adding extra complexity to improve
the accuracy of specific models.

MAX_INTEGRATION_TESTS_REV_ID: 2ec66c63435519c83a2abf549fddd203a2dabea1