[stdlib] Rename `empty` origin to `external and add `ExternalPointer` alias.
This is another step in the larger refactoring work of UnsafePointer as
a whole. This renames the `empty` origin to `external ` to better
capture semantics of what the origin does. This also adds quite a lot of
documentation to the `ExternalPointer ` for guidance on correct usage.

MODULAR_ORIG_COMMIT_REV_ID: 6c2859f4ec11be3eaea35cfd684c054e6cc77d20