[POP] Fix interpreter crash on SIMDXOrOp (#71356)
The first two of the custom folds on SIMDXOrOp were returning a Value
when the interpreter requires an Attribute - i.e., a constant.

For the first fold - `xor(x,0) -> x` - we were unconditionally returning
the operation's LHS operand, not the FoldAdapter's view of that operand
which may have already been folded to an Attribute.

The second `(xor(xor(x,1),1) -> x` fold old only makes sense in a Value
context. This is because it looks through the LHS to see if it was
defined by another XOr which is only possible with a runtime value.

The third fold was the fallback used for all SIMD operations.

To fix the crashes, we reorder the folds to prioritize the fallback that
may return an Attribute. If it does, we choose that. If it doesn't, we
try one of the first two folds. In each case we consistently prefer an
Attribute if we can get one. If not we fall back to the Value-typed
fold. In such cases it's likely okay to return a Value as we'd otherwise
be returning a null Attribute anyway which would crash the interpreter
regardless.

MODULAR_ORIG_COMMIT_REV_ID: 6ff1ca634ccdd0cb9c6568c2613139ebc6e7fcda