[Pipeline] Fix Test for Deepseek Empty Batches to not run during Logit Verification
This test was marked with AITLIB-293 which includes it in logit
verification. As it was specifically merged as a manual broken test to
ease group debugging, I am removing this tag to enable it to be run on
main, outside of CI and Logit Verification.

MAX_INTEGRATION_TESTS_REV_ID: ff2a9ffd85ba8da3f79f4a21d98c8835c54f3497