[Mojo] Remove Incorrect Type Check (#71241)
This type check is not correct when the call function captures
parameters and has parameters. Correcting it is possible by rebinding
correctly but this is already performed by the verification pass so
let's not duplicate it here.

MODULAR_ORIG_COMMIT_REV_ID: 74c333e05c2500a979aee0accd86dad70a734055