[Benchmarking] Elide base64 data URIs in input prompts when using `--print-inputs-and-outputs`
Before the string of a prompt with an image could be half a million
characters long.
This was unprintable. This is useful to identify correctness bugs in
outputs for
benchmarks involving images (eg: QwenVL).

I also include a hash in order to allow programmer to easily notice if
images
are repeated in multiple requests.

MODULAR_ORIG_COMMIT_REV_ID: 648a8f7af0dd94de0fd5bd70b3a5232ab81dcb86