[Scheduler] Use kwargs in BatchInfo class and add `preempted` and `cached_toks` fields
This is nicer to read than the previous version but uses up more column
space.
In order to avoid wrapping lines, we surround the list of expected
batch_infos
with `fmt: off` and `fmt: on` directives to tell `black` to not format
that
section.

The `preempted` and `cached_toks` fields are newly added. `preempted` is
the number of
requests that were preempted while scheduling this batch. `cached_toks`
is the
total number of cached context tokens across all requests in the batch.
The `cached_toks`
field is useful in the subsequent PR to write tests for the
`max-batch-context-length` feature.

We also replace `assert len(actual) == len(expected) and actual ==
expected` with
`assert_batch_info_equal(actual, expected)`. This produces more readable
failures.

MAX_TESTS_TESTS_ORIG_REV_ID: 9ccfe15ba6b6683ad814e036046789bda130e30e