[POP] Introduce a #pop.simd_and attribute (#71611)
This is a compile-time version of the pre-existing `!pop.simd.and`
operation. It can be used to fold `!pop.simd.and` in functions marked
`@always_inline("builtin")`.

This attribute shares folding logic with the operand. It could also be
used as the base for future binary SIMD operators, if and when we move
away from special Int and Index types.

Towards MOCO-2534.

MODULAR_ORIG_COMMIT_REV_ID: a65a543b42b1923c09a090043505afda4815768b