[Serve] Clip Max Sequence Length based on the Max Single Sequence Length in the KV Cache
With longer context models it is not uncommon that we are constrained by
the ability to fit a single sequence at max length in the KV cache. As a
result, when we currently pass a sequence beyond the capacity of the KV
Cache, within the max context length of the model, we error out and fail
on the model worker. This is taking down the entire server for
occasional long requests. To simplify this, I've exposed a method on the
MemoryEstimator that estimates the single maximum sequence length from
the KV
Cache and sets this value appropriately within config resolution.

As a result, when a long context object is now passed we error out
during `new_context` returning an error to the user in the API server,
before the invalid request makes it to the model worker.

MAX_INTEGRATION_TESTS_REV_ID: 7d6e855e748f6b4e2736ff49b19a0804b597ca3d