[AsyncRT][NFC] Extract parts of the CUDADriver to a new class (#71087)
Extract the functions from the CUDADriver used for compilation to a new
CUDACompilationDriver class. This class will _not_ initialize the CUDA
driver, so that we can use it when there is not a physical GPU present
on the machine.

This change should not modify behavior of the code in any way.

This is part of a larger effort to allow cross-compilation of models
which require a GPU to run, but can be compiled on a machine without a
GPU present.

Towards GEX-2483

MODULAR_ORIG_COMMIT_REV_ID: 6e5b91f8dd1ba72a600ded2f7663c703552a3740