[KGEN] Better align syntax of 'emit' options with `mojo` (#71451)
This commit better aligns the syntax of various emission options for the
`kgen` tool with `mojo`.

For example:

* `mojo -emit object` vs `kgen -emit`
* `mojo -emit llvm` vs `kgen -emit-llvm`
* `mojo -emit asm` vs `kgen -emit-asm`
* `mojo -emit shared-lib` vs `kgen -emit=shared` (note: discrepancy)

This caused confusion for users switching between the tools and likely
caused unintentional copy/paste errors in documentation. Now the `kgen`
options syntax are aligned with `mojo`:

* `mojo -emit object` vs `kgen -emit=object`
* `mojo -emit llvm` vs `kgen -emit=llvm`
* `mojo -emit asm` vs `kgen -emit=asm`
* `mojo -emit shared-lib` vs `kgen -emit=shared-lib`

Note however that for `kgen` the style with a space is still not
permitted: `--emit llvm` will fail; the `=` style must be used. Note
also that the old `--emit` (i.e., without `=object`) is still supported
due to a quirk in the options parser. These are both due to an
implementation detail that can be ironed out in future work with a
larger refactor.

MODULAR_ORIG_COMMIT_REV_ID: b18b3a58283834a424ded1865f9adccdefd9046b