[Mojo] Strip EmissionKind From Units in Attributes (#72075)
If a function f is serialized into a package A the emission kind in its
subprogram attribute is FULL. If I create a mojo file that depends on
package A but is built with debug info level LINETABLES, then f will
have the wrong emission type when materialized into the context from the
package. To solve this problem we have a stripDebugInfo pass that
corrects the mismatch but it was only applying this correction to the
locations not the attributes. Since the closure op contains an attribute
with a subprogram attribute the emission kind in this subprogram
attribute was not getting updated.

This PR corrects that behavior, with the intent to follow up later to remove the attribute and revert this change

MODULAR_ORIG_COMMIT_REV_ID: 8e9b175b691235a1fbd79aa5e9794369a9a5c662