[MOGG] Test StaticTensorSpec propagation through fusion with EmitMojo (#71852)
In KGEN, We used to test the correct propagation of the strides after kernel fusion by printing the specs of the fused tensor inside some test kernels.

But I realize this doesn't make much sense, let's stay we fuse an input with a view op, then print the static strides of the `FusedInputTensor` object.
This object is an empty tensor (the data pointer is None). The shape param make sense, but the strides / alignment etc is irrelevant since we can't access the data pointer anyway.

In KGEN we used to build the FusedInputTensor with the specs of the transformed view op.

But in practice we want to test the specs of the view tensor used inside the input lambda has (this tensor isn't passed to the kernel directly.

To accomplish this, the PR adds a new option `moggDebugDumpFusedTensorSpecs` to add debug prints of the fused tensor specs inside the fusion lambdas.
These prints can be tested with the `// CHECK` directives when running the test.

- The propagate-strides.mlir test was updated to check the values of the specs inside the fusion lambdas instead
- The e2e-fusion.mlir test was updated to only check the FusedTensor object specs (the strides propagation tests should be enough to catch bug anyway)
- Add a new test to check the specs of the the fused tensor inside an elementwise body lambda.

Fix GEX-2745

MODULAR_ORIG_COMMIT_REV_ID: 75ef61e5090fc8f75e3884d0f5ad3b2cc4f8843a