[AsyncRT][NFC] Refactor CUDA for compilation only  (#71314)
These changes modify the device context and CUDA so that we can separate
compilation from execution behavior and allow virtual devices to compile
CUDA code without the need to call `cuInit` and have a GPU present on
the machine.

The `CUDADevice` will now forward compilation calls to the
`CUDACompilationDevice`. The virtual device will now use the same
`CUDACompilationDevice` when it needs to compile without a GPU present.

This is part of a larger effort to allow cross-compilation of models
which require a GPU to run, but can be compiled on a machine without a
GPU present.

Towards GEX-2483

MODULAR_ORIG_COMMIT_REV_ID: 37f8047ceabd1d4903d285498b3a5d90238ecd5e