[SDK] Use bias dtype from checkpoint if model dtype is `float8`
This PR fixes a dtype mismatch bug in `Linear`, where the constructor
used the `dtype` parameter for the bias. For `float8` models, we now
capture the bias dtype from the checkpoint and plumb it through the
`Float8Config` so that we can use it in the linear layer.

MODULAR_ORIG_COMMIT_REV_ID: b48111263d4968fd5cc7a01708d18595c916176b