[Models] Move the smoke test model selection logic from yaml to python (#72097)
The logic in the yaml file is getting out-of-hand when adding multi gpu,
so now is an appropriate time to move it over to python.

Note that this PR also adds support for 2xH100 runners, specifically for
the Gemma model.

Note: In the yaml file we do not call the script via bazel, but use pipx
which is installed on all runners.
This is to be practical, pipx is instant and we do not want the model
selection to potentially take minutes

REF PAQ-1488

MAX_INTEGRATION_TESTS_REV_ID: 0f73e7956a3ba1e36127cb38193b1a341055037f