[SDK] Provide a null KV cache manager for compile-only mode  (#71492)
Use the null object pattern to provide a KV cache manager object that
will not do any GPU memory allocation. This is useful during
compile-only mode, as we need to avoid allocating GPU memory when we
don't have a physical GPU device available.

This is part of a larger effort to allow cross-compilation of models
which require a GPU to run, but can be compiled on a machine without a
GPU present.

Towards GEX-2483

MODULAR_ORIG_COMMIT_REV_ID: 6d7aa8c207748cf900d99aea1039b8b2421535fb