[Cache] Do not use llvm::bit_cast to generate cache key (#72161)
Cache keys are used to generate file names for objects in the cache.
The current process for generating the key hash for the
TransformCacheKey involves generating the 128-bit xxhash and then using
`llvm::bit_cast` to reinterpret the internal bytes of the xxhash as
a string.

This produces odd results where the same hash will always fail
to lookup in the cache. My speculation is that some of the bytes of the
hash are not legal path components (e.g. '\0').

This change updates the code to use the raw_ostream wrappers for xxhash
in our codebase, which include a convenience function to get a string
representation of the hash. That generates a hex representation that
should be a legal file name on all platforms.

MODULAR_ORIG_COMMIT_REV_ID: 7403f937c3bd33e9016294d2c732004d5bd5fa85