[LoRA] Remove continuous response worker for LoRA operations
This removes the continuous response worker that polled LoRA responses from the model worker. This was slowing down the frontend and stalling requests even when they weren't interacting with the LoRAQueue. Instead, we are choosing to poll the response and sleep the thread only when we issue a LoRA request for loading/unloading.

This also adds extra bookkeeping to the LoRAQueue to keep track of loaded/unloaded LoRAs so that we don't need to poll the model worker for every request when LoRA is enabled on the server. Previously, for every request, we had to check what LoRAs were loaded and compare with the incoming request to see if it was valid.

MODULAR_ORIG_COMMIT_REV_ID: 0856e223dfbc3a7d0311a9114bb1d973d87f91fc