[LoRA] Speed up base model perf
We add a conditional to check if there are LoRAs in the batch of
requests. If there are no LoRAs, then we take a codepath without the
LoRA computation else we take the path with LoRA computation.

MAX_INTEGRATION_TESTS_REV_ID: a591d977d69c583ef2646d9a5a59beb02515b420