[Mojo] Remove Closure Trait Bodies In Repl (#72183)
We need the traits preserved in the module if functions or structs
reference them because callees will need to bind or rebind their
implementations to traits of this type, which are defined through the
methods. The problem is that in the repl the bodies of trait functions
are deleted, resulting in a verification error when these structures are
preserved.

MODULAR_ORIG_COMMIT_REV_ID: 4be02fa2869bda8b834a38cbef7c4ddec687841f