[mojo-lang] Always print key sugar even if not explaining. (#72440)
If sugar is a key nested value in a type mismatch, make sure to
unpack it, even if not "diffing". This allows us to print the
expansion of `ideal_width` in the b example.

This also renames "original" to "expanded" in SugarAttr which
makes it easier to explain.

MODULAR_ORIG_COMMIT_REV_ID: 1205406cba89348b879579829775f67e659b3687