[KGEN] Move "SIMD Construction Checks" into the compiler (#72343)
This commit starts to move the restrictions that the stdlib places on
SIMD types into the compiler.

SIMD types are special and are inherently understood by the compiler so
it makes sense to have the standard library and compiler aligned on the
constraints placed upon the type. This should also make the compiler
more robust when Mojo is skipped and we test the internals, e.g,. with
MLIR.

One benefit of doing this is that we will be able to remove uses of
`constrained` from SIMD constructors, which are blocking attempts to
mark them `@always_inline("builtin")`.

Towards MOCO-2811.

MODULAR_ORIG_COMMIT_REV_ID: 8b4eedab8c3699f8c540b57a3304bb7e769602dd