[mojo-lang] Prune sugar for aliases that starts with a `_` (#72508)
This changes the compiler to elide sugar for aliases that start
with a `_` to reduce the size of the IR and make it easier to write
testcases.  These aliases are presumably internal implementation
details of types.

This also expands the doc to talk about this.

MODULAR_ORIG_COMMIT_REV_ID: cc9b65dbecfa16aa522ab68de5613ab86892a95d