[Support] Add missing direct deps (#72522)
This adds some missing `deps` in the bazel config where the target only
built because the dep was included transitively. Now everything that the
target imports is directly listed in `deps`. This makes it easier when
someone is removing deps from somewhere else in the graph, and it makes
them less likely to have unrelated breakages because of these implicit
deps. This is also a stepping stone to enable some sort of
include-what-you-use tool where it is easier to know about the existing
headers if deps are strictly listed here.

MODULAR_ORIG_COMMIT_REV_ID: 7bafae867a3c2de71a00888a681cbe05fd3bf314