[KVCache] Rename `maybe_reserve` to `alloc`
This API allocates the blocks needed by a request to run for N steps.
Some of these blocks may be reused from the prefix cache.

As it is possible for there to be insufficient blocks to satisfy the
allocation, we previously returned a bool to indicate success or
failure.

This API was problematic as the error code was largely ignored. For
example,
in our tests you can see calls like:
```
paged_cache.maybe_reserve(ctx, num_steps=1)
```

Notice that the returned boolean value was never checked.

Now the API raises an `InsufficientBlocksError` error instead. This has
a more informative error message than just returning false. In cases
where
we need to gracefully handle out-of-block exception, we simply wrap the
call with a try/except block:

```
try:
    paged_cache.alloc(ctx, num_steps=1)
except InsufficientBlocksError:
    # Handle the error gracefully
    ...
```

I also modify the docs so that all mention of `prefetch` are purged.

MODULAR_ORIG_COMMIT_REV_ID: 55ce59d3d0afada4be6420b182701e01f3fc2582