[MGP] Update mgp.buffer.constant to use new borrow semantics (#72578)
This change moves towards the planned way to treat MGP ops that return
data that, in some way, they do not own. In the case of
`mgp.buffer.constant`, this the memory storage handle that is underlying
the packed data is null, to indicate that the constant data outlives the
use of the data represented by the return value of the operation.

Towards GEX-2381

MODULAR_ORIG_COMMIT_REV_ID: b5db22e88a70433aa25272ff7b62c2c6564f469a