[ExternalFrameworks] Delete it! (#72614)
There is exactly one remaining reference to this folder, which is in
benchmarkPipelineGPURun.yaml. However, the option to run it will never
get hit, as we don't have a config starting with `trt-`, so this is dead
code (as has been suspect for a while).

MODULAR_ORIG_COMMIT_REV_ID: 5644548e1514603e4bc04e971e19927ed00f6d1d