[MOGG] Fix source location handling for bytecode hash (#72467)
The `getBytecodeHash` function currently makes use of a registered
callback to emit all source location attributes as `UnknownLoc`.
In practice, this results in equivalent functions generating different
hashes where the only different is source location information.
When a pre-pass is performed to erase source locations, the functions
hash to the same value.

In the MLIR bytecode format, all attributes are numbered and emitted
into a separate section of the generated bytecode. A pre-pass is
responsible for enumerating all the attributes and assigning a number to
each attribute.

My hypothesis is that while we can override which attributes are emitted
to the bytecode stream, the attribute numbering process still numbers
all the attributes in the IR, even the source locations we don't want to
emit. This results in source locations and other attributes having
different attribute numbers.

This change updates our bytecode hashing infrastructure to use the old
process of: clone op -> erase source locations -> hash op.
This is more expensive but results in a net reduction of stubs generated
from the HoistHarness transform (from 205 to 136 stubs).

MODULAR_ORIG_COMMIT_REV_ID: ebfaf4379b8afd58b4e02d4a45a08e7beb8c0aa8