[Benchmarking] Move benchmark_pipeline_latency.py configs to YAML format (#72665)
This is the first of many steps in trying to port
`benchmark_pipeline_latency.py` to be our POC candidate of `pydantic`
types and a CLI. Should be a no-op to end users.

We'll still merge this anyway to keep it all consistent with the rest of
the benchmark entrypoints, but it also serves as a test for what other
functionalities we need to account for when going through tool
selection. Given this script is relatively simple as its current
implementation, it's a lot easier to roll it out to it first.

Refs PAQ-1637, MAXSERV-1035

## Future Work
1. Make these dataclasses `pydantic` `BaseModel`s.
2.
[Internal link]
- some weird quirks in this script that could use some consolidation /
consistency with our other benchmarking scripts.

## Tests

Benchmark Pipelines GPU -
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 9cdaf57d14dd09625f0f95a0bf8bf510219b86f0