[mojo-stdlib] Remove all the overloads of keep, NFC
I'm not sure the archeology here, but we have a ton of overloads
of `keep` for various special cases.  It is enough to copy the
"kept" value into a var, and pass the address of that var to an
inline asm as a register value.

I ran into this with an unrelated change to overload resolution that
started making it so the SIMD overload got picked... which uses a
weird inline asm, which caused the AMD GPU backend to crash due
to (its) bug not handling the mem/reg register constraint correctly.

MODULAR_ORIG_COMMIT_REV_ID: ceb2a7c85ad75b5b870e50e96a941c38d0443a04