[KGEN] Fix elaborator crash and hang when erroring 💀  (#72795)
When elaborating a ParamNode `p` has errored, the elaborator marks the
node's `paramCh` value to be done so as to wake up the other ParamNodes
that are blocked by `p`.

```
void ParamNode::setToError() {
  if (done.exchange(DoneState::ERROR) == DoneState::NOT_DONE)
    paramCh.copy().emplace();
}
```
However, this operation here doesn't change `p.state` which will still
be in `IN_PROGRESS` state. This is bad, because imagine `b` is the one
that's blocked by `p` (means `b` is the caller who calls `p` as the
called in a kgen.param.apply op). When `p` unblocks `b`, `b` resumes its
own elaboration which will try to `specializeGenerator` of `p` again
(because it was blocked previous to wait for `p` to finish). However,
`p` is still in `IN_PROGRESS` mode, so `b` here will again block itself
to wait for `p` to "finish" and to unblock `b`. We know now that `p` has
already `emplaced` its `paramCh`, so it will immediately unblocks `b`
and continue the infinite loop of blocking and unblocking `b` 💀 .

The root cause of this problem here is somehow we have two variables in
`ParamNode` that both bear the information of its done-ness:
`ParamNodeState state` and `std::atomic<uint8_t> done` (why????????) and
they are not being used consistently.

The fix includes the following changes:

- [x] Mark `ParamNode` to be `DONE` with its `done` when `setToError()`
so that we don't get into the infinite loop of blocking and unblocking
its caller in `specializeGenerator`.
- [x] The tracking of number of work items in ExpanionGraph (i.e.
`g.numWorkItems`) doesn't include errored out node and its caller
dependencies (because the dependencies should be erroring out also).
This means that `g.worklistCh` can be ready when all active work items
are done while the errored out items may still running (because they are
not being tracked by `g.numWorkItems`. We shouldn't start to destroy
nodes right after `g.worklistCh` is done because the destruction may run
in parallel with the on going errored out node.
- [x] All on going blocked tasks are being tracked with
`ExpansionGraph::numOutstandingResources` (another why???? a different
variable to keep track things, why???), and `ExpansionGraph` will wait
for all of them to be done before destructing itself. We can put
releasing `ImplNode::inst` at the same place if elaboration errored so
that we know everyone is truly ALL DONE!
- [x] Because errored out ParamNodes and their blockers are not being
tracked with `g.numWorkItems`, when we complete them, don't add them to
the worklist and then remove (signalWorklist), because that can make
`g.numWorkItems` reach `0` multiple times which leads to `g.worklistCh`
being emplaced multiple times illegally.

MODULAR_ORIG_COMMIT_REV_ID: 6125a4f4df052467de7b55ff09e7b3c4dc7df1fd