[mojo-lang] Struct extensions naming/importing changes (#72612)
Re-merges the compiler parts of
[Internal link] which were always fine
but caught up in the revert of the stdlib changes.

This PR does three main changes:

* Cleanup: Makes extensions (eg extension MyStruct) known to their
parents not as their struct's name ("MyStruct") but instead prefixed
with "extension:" ("extension:MyStruct"). This makes it so
lookupAndResolveDecl("MyStruct") doesn't return the extensions as well.
This is good because we generally like finding and resolving as little
as possible.
* Makes extensions (eg extension MyStruct) also known to their parents
as "extension:". This is helpful when we want to look up "all extensions
in a given scope", which is useful for the next point.
* Makes it so when we import anything from a module, we also import all
extensions (we just look for anything named "extension:").

Note that this might not be the final design, this makes an assumption
that if a file wants to use an extension, then that extension is in a
file that is indirectly imported. It's similar to Rust/Swift/C#/Kotlin,
but we might want to be more similar to Python, which just makes all
extensions globally visible to everyone everywhere, despite all the
spooky action at a distance that entails (doc incoming on this).

MODULAR_ORIG_COMMIT_REV_ID: ff5f6def0da9f9f5c0ea73e3bdd9fc84b7c1437f