[bazel] Remove cc_autolink_deps (#72938)
This rule was a hack because it didn't propagate all private fields from
the underlying library that we are re-exporting with alwayslink. This
was the last use case and we can force the necessary symbol to be linked
with a linker flag instead.

MODULAR_ORIG_COMMIT_REV_ID: e769ae3d96a7d6d8d919229b0c5d61cf44c45f98