[mojo-lang] Teach witness folding to be sugar aware (#72961)
This teaches witness folding to not be confused by sugar. This took
me forever to track down, and was causing a patch to blow out the size
of the generated IR in the presence of more sugar.

MODULAR_ORIG_COMMIT_REV_ID: e5a98689c05a4085944617c57778289a9a4a7bf5