[KGEN] Enhance `pop.cast` index-type folding (#72978)
This commit fixes several issues in the folding of the `pop.cast`
op/attribute when casting from an index type to another.

When casting from an index type to another type (int or floating-point,
in this commit), we can fold in more cases if we know the target's index
bitwidth.

Furthermore, the generic index-type folding logic ensures that the fold
result is identical whether the target's index type is either 32- or
64-bit. If we know both the index type and the resulting type's
bitwidth, we can skip this and fold in yet more cases.

This infrastructure might benefit from a refactor at some point. For
example, it doesn't gracefully handle other result-type bitwidths which
we may be interested in. It's also quite verbose and involves
duplication across multiple cases.

This is a necessary prerequisite to enable the folding of SIMD
constructors as part of MOCO-2811.

MODULAR_ORIG_COMMIT_REV_ID: aee7d3905c6239440d6fd116418ba939d6a3fb58