[KGEN] Fix invalid source locs for postfix unary exprs (#72577)
This commit fixes invalid source locations for postfix unary
expressions, where the start of the source range started *after* the end
of the source range. This led to all sorts of undesirable behaviour when
printing diagnostics, such as compiler crashes or memory corruptions.

The `UnaryOpNode` took a location for the beginning of the expression,
and used the sub-expression's location as the end. This works find for
regualr or prefix unary expressions but not for postfix ones where that
sub-expression started before the newly parsed token.

This commit is a simple fix which adds an explicitly tracked ending
location, and overloads the constructor to optionally take that explicit
location when constructing a postfix transfer expression.

Fixes MOCO-2698.

MODULAR_ORIG_COMMIT_REV_ID: 35f4372f96fb70214bd0f56490a6fa5586081e8d