[Benchmarking] Add GPU kernel launch benchmark using CUDA (#73114)
Adds C++ benchmark which measures the time to enqueue GPU kernels in
CUDA directly. Benchmark is written using the MicroBenchmarks.h
framework and enqueues a batch of 100 kernels each iteration.

This will provide a baseline for GPU kernel enqueue time on Nvidia GPUs
that can be used to compare with comparible benchmarks at other levels
of the stack.

Includes a readme with instructions for how to run, and a common CUDA
header file so that it can be used for future benchmarks.

Closes GEX-2853
MODULAR_ORIG_COMMIT_REV_ID: 884deef9046d72a74c244396ae04d5dfcc2ab2ee