[Models] Change prepare_initial_token_inputs to support data parallelism
Another day, another re-write the world PR.

This method prepares the inputs to the graph for multi-step execution.

Here is the [signature for
[Internal link]

```
    def prepare_initial_token_inputs(
        self,
        context_batch: Sequence[TextContext],
        kv_cache_inputs: KVCacheInputs | None = None,
        return_n_logits: int = 1,
    ) -> DeepseekV3Inputs:
```

Notice that context_batch has type Sequence[TextContext]. This is a
problem for data parallelism. This causes the implementation of above to
need to do the following:

```
    data_parallel_splits: Tensor
    if self.pipeline_config.model_config.data_parallel_degree > 1:
        assert isinstance(self.kv_manager, PagedKVCacheManager)
        data_parallel_splits = self.kv_manager.get_data_parallel_splits(
            context_batch
        )
    else:
        data_parallel_splits = Tensor.from_numpy(
            np.array([0, len(context_batch)], dtype=np.int64)
        )
```

This essentially converts a `Sequence[TextContext]` into
`Sequence[Sequence[TextContext]]` to partition the request by their DP
rank.

We should NOT be doing this. The Scheduler passes TextGenerationInputs
to Pipelines which has the 2d structure already. It makes no sense that
the Pipelines flattens it to 1d and then unflattens to 2d.

Instead we should update the signature of these methods to accept
`Sequence[Sequence[TextContext]]`.
```
    def prepare_initial_token_inputs(
        self,
        batches: Sequence[Sequence[TextContext]],
        kv_cache_inputs: KVCacheInputs | None = None,
        return_n_logits: int = 1,
    ) -> DeepseekV3Inputs:
```

This has a major benefit that we no longer need to call into the KVCache
here. In a future work, I hope to delete
`kv_manager.get_data_parallel_splits(..)` api entirely.

To do this, we need to modify the signature/impl/usage across all
models. For models that do not support DP, we should throw an exception
if `len(batches) > 1`.

In making this change, I created a `flatten2d` util in
`max.support.algorithm`.

MAX_INTEGRATION_TESTS_REV_ID: 1e1c0dc246403568b026afd5e851aad80d72c47e