[KGEN] Compare high level param operators generically (#73327)
Towards MOCO-2880. This PR fixes the first case.

This was not a problem with `where` clauses or `and`, but with our
parameter expression comparator. The simplifier for `and` relied on the
comparator to deduplicate items, and the comparator relied on each kind
of parameter operator attr to implement `isLessThan`. This was not
enforced anywhere, which led to un-comparable parameter expressions.

Instead of implementing it one by one for the new variadic operators,
this PR implements a generic lexicographical comparison for parameter
operators so that non-leaf nodes can transparently be compared via its
operands. The only nodes that will need to implement this `isLessThan`
method going forward are the ones that are leaf nodes (e.g.
ParamDeclRefAttr), or any special operators that can be more efficiently
or more correctly implemented in a custom way.

MODULAR_ORIG_COMMIT_REV_ID: c1874c1e7c10f25e7ad9c5298233f33d2e69e3ff