[stdlib] polish: Remove ambiguous `String.join` variadic overload (Part 2.5/N)
This avoids a change in overload resolution behavior that will occur
when making `List` conform to `Writable`.

Code such as `"foo".join([1, 2, 3])` ought to resolve to the
`String.join(Span)` overload, but could theoretically also
resolve to this removed variadic overload as a single-element
variadic argument pack. That does not occur today because
the variadic overload requires the elements to
implement `Writable`, which `List` does not _today_.

To avoid a bug in a subsequent PR that _does_ add a `List(Writable)`
conformance, simply remove this overload.

This is directionally right longer term, as Python does not have a
variadic `str.join()` method, only a method that takes an `Iterable`.

MODULAR_ORIG_COMMIT_REV_ID: 37a4c8fbefef73ca894f6eb4d17974fe3900fc13