[Context] Rename maybe_chunk to chunk and raise
Previously, we were returning the number of tokens returned in the
active_length after chunking. Practically, this is not how this value is
used. Either the value provided is invalid, and we cannot chunk, or we
chunk appropriately. As such, this will now error out. This forces the
callers directly to manage the error, which should only occur as a logic
error in implementation.

MAX_INTEGRATION_TESTS_REV_ID: 77d34b1fbbe200f0529be1592176a3565fadbbb9