[AsyncRT][CUDA] Refactor CUDA driver loader (#73364)
Currently to add a new CUDA API to our loader you would have to:

* Add the type and a member for the function pointer
* Add a wrapper without the `cu` to log and call the function pointer
* Add a `GET_FUNC` entry to do the `dlsym`

This patch refactors the loader so that all of these are generated by
macros and we can have a single file listing the CUDA APIs we want to
support, meaning that adding a new API is now just about adding a single
entry in `CUDADriver.inc`.

Additionally it renames the wrapper functions to match the exact CUDA
functions they're calling, which makes it easier to search for specific
CUDA calls, slightly clearer about when we're calling in to CUDA, and it
makes more sense with the macro approach.

MODULAR_ORIG_COMMIT_REV_ID: bade1b5f3e6b329c3c8ee5abe3c4ce1914fc0bcd