[mojo-lang] Ensure shift folding is conservative if target is not available (#73416)
I.e. when the target is not known, or the bit width is not 32 or 64, we
only fold if the result of the latter two match. The patch also
refactors a helper that will be used later in other folders.

MODULAR_ORIG_COMMIT_REV_ID: bf0454f1c6efbb24eff6cdc5a2e6881420e1dfea