[Scheduler] Simplify advancing requests in batch constructor
This moves the responsibility of managing the responses object inside
the Scheduler. The batch constructor should not be responsible for
mutating the responses object itself. It should just be responsible for
tracking batching state, and creating new batches. It is the schedulers
responsibility to ensure the correct resposes are returned to the API
worker.

MODULAR_ORIG_COMMIT_REV_ID: a4159f272f50aaa99311ba96446fbd04adf58561