[Mojo] Remove Closure Wrappers (#73474)
The key of the closure wrapper cache was not accounting for the trait
type of the struct wrapper which renders it unique. This PR fixes that
issue by replacing the key with the trait type then mangling the name of
the struct from this trait type to guarantee uniqueness. This is a
requirement before adding implicit conversion of closures to function
pointers.

MODULAR_ORIG_COMMIT_REV_ID: 664b3210c2725462d903cdb8415d5b7d516140e2