[KGEN] Make sure interpreter get type size that is correct with how it is used.  (#73487)
The interpret stores and loads attributes to its own memory. Some of
them `TypeType` and `StringType`, they are always stored as 64bit
integer numbers for the pointer and size. This makes sense because the
pointer can be an opaque pointer of the mlir attribute in the compiler
that most of time, the pointer is 64bit for most of the host systems.
When we store those values, we need to make sure that we create an
`APInt` of 64bit instead of the ptrSize of a specific target which can
be 32bit.

MODULAR_ORIG_COMMIT_REV_ID: a468c1c59c142bead250ccbd82e7aa3d287174bf