Fix Deque shrink realloc bug when capacity > min_capacity
When _head == _tail, the deque is either empty or completely full.
The previous code always used _capacity for deque_len in this case,
which was wrong when shrinking an empty deque after pop operations.

This fix uses _capacity only when growing (buffer is full), and
len(self) when shrinking (buffer is empty).

Fixes #5635

Signed-off-by: magi8101 <mj7841@srmist.edu.in>