[LIT] Use the element type to bind VariadicPack (#74160)
Since `VariadicPack` uses `posvaarg` convention, meaning that we need to
unpack the variadic type and use the element type to match the
VariadicType.

This is a deeply hidden bug that was revealed when re-working on
parameter inference. It somehow did not cause any issue (the failure was
recorded, but somehow not reported). But it does lead to error when I
made the switch to the new linear inference algorithm.

Towards MOCO-2940

MODULAR_ORIG_COMMIT_REV_ID: dbaf7bd19d07a0f83a9c874f25c5d2627ea321fa