[Support] Apply misc-include-cleaner checks (#74088)
Whether or not we actually enforce these checks is a different
discussion, but I think the results themselves should be
uncontroversial.

This is a best-effort pass, there's a whole bunch of changes I had to
revert but this is probably 90% of the changes.

MODULAR_ORIG_COMMIT_REV_ID: f1c19707f8eeeb7224446c8d0b9b647cbda25fe0