[Interfaces] New Token Management APIs
Managing tokens directly inside context objects, leads to very
complicated internal logic, and a lot of code. I have seperated out the
core logic for interacting with an inflight token array and moved it
into max.interfaces. This is a highly tested class, which should offer
all the APIs needed for interacting with a token array throughout MAX
downstream.

Few important details to call out:

- `chunk` API introduced. This API takes in a number of tokens to chunk,
and chunks the token buffer. If the chunk size cannot be provided, this
will raise an error.
- `TokenSlice` TypeAlias has been introduced, this allows us to
explictly identify that we are expecting a (sub)section of a token
array.
- `consume_recently_generated_tokens` more explictly calls out side
effecting behaviour identifying that this method returns and advances
internal state, such that the outstanding completion tokens are not
retrieved more than once.
- A few additional errors are raised regarding invalid state, I hope
this can catch errors early regarding invalid use of the API. This is
likely to cause integration pain in the short term.
- This is not meant to replace EOS or status tracking, I expect that to
remain within the Context objects.
- This does change the behaviour slightly with regard to completion
responses with EOS. We now return the EOS token, where previously we did
not. Note this gets removed during detokenization anyways, and ensures
that the behaviour is more consistent for APIs such as completions which
expect a detokenized response is returned in full.

MODULAR_ORIG_COMMIT_REV_ID: 3132e3fd33936eb44fb8a1927352b62f9894ccdb