[mojo-lang] Relax assertion in pop.cmp folding utility (#74255)
The `cmpOpfoldHelper` has special folding rules for inequalities that
compare a value to 0. These assumed that the compared value is not
constant, but this assumption can be violated if the operation doesn't
yield the same result on 64 and 32 bit systems (and the target is not
known. This patch relaxes this assumption from an assert to a fold
failure.

MODULAR_ORIG_COMMIT_REV_ID: 053cbc4582c7759587c3cca609c139e2cb17dd77