[MAX] Allow output tensor to be allocated on pinned memory for ParallelArrayOps.concatenate
This PR allows user to pass an accelerator to
ParallelArrayOps.concatenate
so that the result may be allocated on a pinned buffer on the specified
accelerator.

It is not guaranteed that the result will be allocated on a pinned
buffer.
For example, if the user passes in a single array to concat, we will
just
return the original array if default_copy is False.

Now the result is a MAX tensor instead of a numpy array.

This PR solely brings up functionality and verifies that it works in
`test_parallel_concat`.

The subsequent PR will make it performant so we don't add this feature
to `test_benchmark_parallel_concat` yet. I am breaking it into smaller
PRs to make it easier to review.

MODULAR_ORIG_COMMIT_REV_ID: 81925d5d74bc206e483894260399132e33fd6129