[KGEN] Fix infinite error emitting logic goof when printing out codegen_unreachable errors.  (#74188)
- [x] We try to print out the call instantiation path from the
callInstantiationGraph recursively from child to parent where there is a
codegen reachable violation. But this graph can naturally have cycles in
it for functions that will call recursively at runtime (which is totally
legal as long as there is a leaf case to break the cycle at runtime). We
need to stop emitting error when we see a cycle in the graph so stop
infinite recursion.
- [x] Update test to cover recursion case.

MODULAR_ORIG_COMMIT_REV_ID: 0fd2e5cc474d65e68d182dcf25f7a882893dd662